+++ run_tests.py
@@ -235,9 +235,15 @@ def main():
         println('Running tests...\n', STATUS)
 
         for test_class in get_integ_tests(args.specific_test):
-          run_result = _run_test(args, test_class, output_filters)
-          test.output.print_logging(logging_buffer)
-          skipped_tests += len(getattr(run_result, 'skipped', []))
+          if integ_runner.assert_tor_is_running():
+            run_result = _run_test(args, test_class, output_filters)
+            test.output.print_logging(logging_buffer)
+            skipped_tests += len(getattr(run_result, 'skipped', []))
+          else:
+            # our tor process died
+
+            error_tracker.register_error()
+            break
       except KeyboardInterrupt:
         println('  aborted starting tor: keyboard interrupt\n', ERROR)
         break